+++ packages/core-js/internals/object-prototype-accessors-forced.js
@@ -2,9 +2,14 @@
 var IS_PURE = require('../internals/is-pure');
 var global = require('../internals/global');
 var fails = require('../internals/fails');
+var userAgent = require('../internals/engine-user-agent');
 
 // Forced replacement object prototype accessors methods
 module.exports = IS_PURE || !fails(function () {
+  // This feature detection crashes old WebKit
+  // https://github.com/zloirock/core-js/issues/232
+  var webkit = userAgent.match(/AppleWebKit\/(\d+)\./);
+  if (webkit && +webkit[1] < 535) return;
   var key = Math.random();
   // In FF throws only define methods
   // eslint-disable-next-line no-undef, no-useless-call -- required for testing